requirements.txt
PyYAML
click
boto3